requirements.txt
Pillow==9.5.0
Flask==2.3.2